Installation an Namen Kivitendo angepasst, Parameter f --fast hintzgefügt

+++ scripts/remove.sh
@@ -9,14 +9,20 @@ echo "Deinstallation LxCars."
 echo "Verzeichnis CRM: $DIR_CRM"
 echo "Verzeichnis ERP: $DIR_ERP"
 echo "*******************************************"
-/usr/bin/sudo -u postgres dropdb lxcars
-echo "Datenbankuser www-data wird gelöscht."
-echo "*******************************************"
-/usr/bin/sudo -u postgres dropuser www-data
-echo "*******************************************"
+
+if [ "$1" = "-f" ]; then  ## für fast
+	echo "Datenbank lxcars und Datenbankuser www-data werden NICHT gelöscht."
+else
+    echo "Datenbank lxcars wird nun gelöscht."	
+	/usr/bin/sudo -u postgres dropdb lxcars
+	echo "*******************************************"
+	echo "Datenbankuser www-data wird gelöscht."
+	/usr/bin/sudo -u postgres dropuser www-data
+	echo "*******************************************"
+fi
 
 #Links löschen und *.orig in * umbenennen
-## Menü löschen
+## Menü Patch rückgängig machen
 if [ -f $DIR_ERP/menu.ini.orig ]; then
    rm $DIR_ERP/menu.ini
    mv $DIR_ERP/menu.ini.orig $DIR_ERP/menu.ini
@@ -24,19 +30,13 @@ else
    echo "Error menue.ini.orig nicht gefunden"
 fi
  
-#if [ -f $DIR_CRM/tpl/firmen3.tpl.orig ]; then
-#  rm $DIR_CRM/tpl/firmen3.tpl
-#  mv $DIR_CRM/tpl/firmen3.tpl.orig $DIR_CRM/tpl/firmen3.tpl
-#else
-#  echo "Error firmen3.tpl.orig nicht gefunden"
-#fi 
 
 ## Bild löschen 
-if [ -f $DIR_ERP/image/lx-office-erp.png.orig ]; then
-   rm $DIR_ERP/image/lx-office-erp.png  
-   mv $DIR_ERP/image/lx-office-erp.png.orig $DIR_ERP/image/lx-office-erp.png
+if [ -f $DIR_ERP/image/kivitendo.png.orig ]; then
+   rm $DIR_ERP/image/kivitendo.png  
+   mv $DIR_ERP/image/kivitendo.png.orig $DIR_ERP/image/kivitendo.png
 else
-   echo "Error lx-office-erp.png.orig nicht gefunden"
+   echo "Error kivitendo.png.orig nicht gefunden"
 fi 
  
 ## Icons löschen
@@ -64,13 +64,7 @@ else
    echo "Error firma1.tpl.orig nicht gefunden"
 fi 
 
-## Patch für stdlib rückgängig machen
-if [ -f $DIR_CRM/inc/stdLib.php.orig ]; then
-   rm $DIR_CRM/inc/stdLib.php
-   mv $DIR_CRM/inc/stdLib.php.orig $DIR_CRM/inc/stdLib.php
-else
-   echo "Error stdLib.php.orig nicht gefunden"
-fi 
+
 
 
 echo "done!!"
